Problem: readdir_r deprecated in glibc 2.24
Solution: use readdir and protect the calls to it with a global mutex.
This is necessary because readdir has not yet been made thread safe.
It might happen in the next version of POSIX, but until then we must
ensure 2 threads never call readdir on the same directory, so a mutex
is necessary.